Updated the project to save outcomes in a manner more consistent with FCModel notation.  Modified the win outcomes from boolean to a new enumeration.

+++ PaperScissorsRock/ConnectionsViewController.m
@@ -134,7 +134,7 @@
         }
         else if (state == MCSessionStateNotConnected){
             if ([_arrConnectedDevices count] > 0) {
-                int indexOfPeer = [_arrConnectedDevices indexOfObject:peerDisplayName];
+                NSUInteger indexOfPeer = [_arrConnectedDevices indexOfObject:peerDisplayName];
                 [_arrConnectedDevices removeObjectAtIndex:indexOfPeer];
             }
         }
